=== Refactor: renames hocs ===
