+++ src/components/hocs/with-active-flag/with-active-flag.jsx
