+++ src/components/hocs/with-load-flag/with-load-flag.test.js
