+++ src/components/offers-list/offers-list.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 
 import OfferCard from "../offer-card/offer-card";
 
-import {withLoadFlag} from "../hocs/withLoadFlag/with-load-flag";
+import {withLoadFlag} from "../hocs/with-load-flag/with-load-flag";
 
 const OffersList = (props) => {
   const {offers, cardClass} = props;
